pyproject.toml
[tool.poetry]
name = "poetrytest"
version = "0.1.0"
description = ""
authors = ["leon16c <leon16c03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cinemagoer = "^2022.12.27"
aiohttp = "^3.8.3"
imdbpie-async = "^1.0.0"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
